# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Packing Stacking API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/packs/add": {
            "get": {
                "description": "Add a new pack with unique quantity",
                "summary": "Add pack",
                "operationId": "add-pack",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/packs/calculate": {
            "get": {
                "description": "Calculates the needed packs combination for a given quantity",
                "summary": "Calculate packs",
                "operationId": "calculate-packs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/packs/list": {
            "get": {
                "description": "List all available packs",
                "summary": "List packs",
                "operationId": "list-packs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/domain.Pack"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packs/remove": {
            "get": {
                "description": "Removes a pack with a certain quantity",
                "summary": "Remove pack",
                "operationId": "remove-pack",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Pack": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                }
            }
        }
    }
}